Support 64-bit windows builds
Added support for 64-bit version of compiler:
- MSVC 2008 project files
- x64 dependencies
- warning fixes

+++ src/bspfile_ibsp.c
@@ -551,7 +551,7 @@ void WriteIBSPFile( const char *filename )
 	/* add marker lump */
 	time( &t );
 	sprintf( marker, Q3MAP_MARKER, Q3MAP_VERSION, asctime( localtime( &t ) ) );
-	AddLump( file, (bspHeader_t*) header, 0, marker, strlen( marker ) + 1 );
+	AddLump( file, (bspHeader_t*) header, 0, marker, (int)strlen( marker ) + 1 );
 
 	/* add lumps */
 	AddLump( file, (bspHeader_t*) header, LUMP_SHADERS, bspShaders, numBSPShaders * sizeof( bspShader_t ) );
